Prevent possible infinite loop.
updateFrame() can potentially call setState, causing
componentDidUpdate() to fire again, calling updateFrame(),
which can call setState() yet again and blow the stack.

We instead only updateFrame() after an update if props have
changed, and one additional time at boot to fill out
the container after mount.

To ensure containers are fully filled, we allow updateFrame()
after a `size` change, but we ensure that not too many of them
are able to fire in sequence to prevent a stack overflow
or infinite loop.